// src/letters/dto/letters.dto.ts
export class LettersDto {
    id?: number;
    companyCode: string;
    unitCode: string;
    staffId?: string;
    offerLetter?: string;
    resignationLetter?: string;
    terminationLetter?: string;
    appointmentLetter?: string;
    leaveFormat?: string;
    relievingLetter?: string;
    experienceLetter?: string;
}